Package Deprecation - Import deprecation information into catalog (#542)
